# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * VidSnap Video Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeInput {
  /** Video URL to analyze */
  url: string;
}