# lib/api-zod/src/generated/types/editInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * VidSnap Video Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { EditInputOutputFormat } from './editInputOutputFormat';
import type { EditInputVoiceEffect } from './editInputVoiceEffect';

export interface EditInput {
  /** ID of a completed download job to edit */
  jobId: string;
  /**
     * Start time in seconds for trim
     * @nullable
     */
  trimStart?: number | null;
  /**
     * End time in seconds for trim
     * @nullable
     */
  trimEnd?: number | null;
  /** Attempt to remove watermark */
  removeWatermark?: boolean;
  outputFormat?: EditInputOutputFormat;
  /**
     * Playback speed multiplier (0.25-4)
     * @minimum 0.25
     * @maximum 4
     * @nullable
     */
  speed?: number | null;
  /**
     * Remove audio track from output
     * @nullable
     */
  muteAudio?: boolean | null;
  /**
     * Volume percentage (0-300)
     * @minimum 0
     * @maximum 300
     * @nullable
     */
  volume?: number | null;
  /**
     * AI voice effect preset
     * @nullable
     */
  voiceEffect?: EditInputVoiceEffect;
}