# lib/api-zod/src/generated/types/platformList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * VidSnap Video Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { Platform } from './platform';

export interface PlatformList {
  platforms: Platform[];
}