# lib/api-zod/src/generated/types/videoInfo.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * VidSnap Video Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoFormat } from './videoFormat';

export interface VideoInfo {
  url: string;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  /**
     * Duration in seconds
     * @nullable
     */
  duration?: number | null;
  /** e.g. YouTube, TikTok, Facebook */
  platform: string;
  /** @nullable */
  uploader?: string | null;
  /** @nullable */
  viewCount?: number | null;
  /** @nullable */
  description?: string | null;
  formats: VideoFormat[];
  /** How the info was extracted */
  sourceType?: 'ytdlp' | 'hls' | 'browser';
  /** Whether the video is a live stream */
  isLive?: boolean;
  /** Whether video is age-restricted */
  ageRestricted?: boolean;
  /** Whether a login/cookie is required to download */
  requiresLogin?: boolean;
}